refactor: Make all dataloaders inherit BaseDataLoader directly
- Move the shared feed engine (prepare, from_*, extraction) into BaseDataLoader
- Move the schema builders and metadata derivation into the base schema module
- Reduce SoccerDataLoader to its download and manifest specifics
- Make DummySoccerDataLoader a direct BaseDataLoader subclass under _soccer
- Validate the config dataloader class against BaseDataLoader in the CLI
