+++ src/sportsbet/datasets/__init__.py
@@ -9,8 +9,8 @@ from ._base._schema import (
     optional_col,
     required_col,
 )
-from ._dummy import DummySoccerDataLoader
 from ._soccer._dataloader import SoccerDataLoader
+from ._soccer._dummy import DummySoccerDataLoader
 
 __all__: list[str] = [
     'BaseDataLoader',
