+++ src/sportsbet/datasets/_base/_schema.py
@@ -1,10 +1,25 @@
 """Schemas for validating statistics and odds data."""
 
-from typing import Any, Self
+from typing import Annotated, Any, Self
 
 import pandas as pd
 import pandera.pandas as pa
 from pandera.typing.pandas import Timedelta
+
+# Columns that identify a match snapshot and place it in time.
+IDENTITY_COLS = ['date', 'league', 'division', 'year', 'home_team', 'away_team']
+EVENT_COLS = ['event_status', 'event_time']
+# The possible event statuses, ordered.
+STATUSES = ['preplay', 'inplay', 'postplay']
+# Types of the identity columns, used to build derived schemas.
+IDENTITY_FIELDS = {
+    'date': Annotated[pd.DatetimeTZDtype, 'ns', 'utc'],
+    'league': str,
+    'division': int,
+    'year': int,
+    'home_team': str,
+    'away_team': str,
+}
 
 
 def required_col(alias: str | None = None) -> Any:  # noqa: ANN401
@@ -100,3 +115,86 @@ class BaseOddsSchema(BaseSchema):
         out = pd.Series(True, index=df.index)
         out.loc[is_post] = ok_post
         return out
+
+
+def odds_columns(columns: list[str]) -> list[str]:
+    """Return the odds columns of a wide frame (the ``{provider}__{market}`` ones)."""
+    return [col for col in columns if '__' in col]
+
+
+def parse_odds_column(col: str) -> tuple[str, str]:
+    """Split a ``{provider}__{market}`` odds column into its provider and market."""
+    provider, market = col.split('__', maxsplit=1)
+    return provider, market
+
+
+def derive_metadata(data: pd.DataFrame, value_cols: list[str]) -> dict[str, dict[str, Any]]:
+    """Derive per-column `include`/`fixed`/`type` metadata from a long snapshot frame.
+
+    Nothing is assumed about a column's role: `include` is the set of statuses at
+    which the column actually carries values, and `fixed` is whether the column is
+    constant within every match.
+
+    Args:
+        data:
+            A long snapshot frame with `event_status` and identity columns.
+        value_cols:
+            The value columns to describe (non-identity, non-event).
+
+    Returns:
+        Mapping of column to `{'type', 'include', 'fixed'}`.
+    """
+
+    def _is_constant(values: pd.Series) -> bool:
+        non_null = values.dropna()
+        return non_null.empty or bool(non_null.min() == non_null.max())
+
+    grouped = data.groupby(IDENTITY_COLS, dropna=False)
+    metadata = {}
+    for col in value_cols:
+        include = [status for status in STATUSES if data.loc[data['event_status'] == status, col].notna().any()]
+        fixed = bool(grouped[col].apply(_is_constant).all())
+        col_type = int if pd.api.types.is_integer_dtype(data[col]) else float
+        metadata[col] = {'type': col_type, 'include': include, 'fixed': fixed}
+    return metadata
+
+
+def _field_name(col: str) -> str:
+    """Turn a column name into a valid Python identifier (``over_2.5`` -> ``over_2_5``)."""
+    return col.replace('.', '_')
+
+
+def _value_namespace(metadata: dict[str, dict[str, Any]]) -> tuple[dict, dict]:
+    """Build the annotations and fields for the value columns from their metadata."""
+    annotations: dict = {}
+    namespace: dict = {}
+    for col, meta in metadata.items():
+        field = _field_name(col)
+        annotations[field] = meta['type']
+        alias = col if field != col else None
+        namespace[field] = optional_col(meta['include'], fixed=meta['fixed'], alias=alias)
+    return annotations, namespace
+
+
+def build_stats_schema(metadata: dict[str, dict[str, Any]]) -> type[BaseStatsSchema]:
+    """Build a statistics schema from the derived value-column metadata."""
+    annotations: dict = dict(IDENTITY_FIELDS)
+    namespace: dict = {col: required_col() for col in IDENTITY_FIELDS}
+    value_annotations, value_namespace = _value_namespace(metadata)
+    annotations.update(value_annotations)
+    namespace.update(value_namespace)
+    namespace['__annotations__'] = annotations
+    return type('StatsSchema', (BaseStatsSchema,), namespace)
+
+
+def build_odds_schema(metadata: dict[str, dict[str, Any]]) -> type[BaseOddsSchema]:
+    """Build an odds schema from the derived market-column metadata."""
+    annotations: dict = dict(IDENTITY_FIELDS)
+    namespace: dict = {col: required_col() for col in IDENTITY_FIELDS}
+    annotations['provider'] = str
+    namespace['provider'] = optional_col(['preplay'], fixed=True)
+    value_annotations, value_namespace = _value_namespace(metadata)
+    annotations.update(value_annotations)
+    namespace.update(value_namespace)
+    namespace['__annotations__'] = annotations
+    return type('OddsSchema', (BaseOddsSchema,), namespace)
